+++ mocks/Application.go
@@ -189,6 +189,24 @@ func (_m *Application) RegisterServiceProviders() error {
 	return r0
 }
 
+// RegisterWithDependencies provides a mock function with no fields
+func (_m *Application) RegisterWithDependencies() error {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for RegisterWithDependencies")
+	}
+
+	var r0 error
+	if rf, ok := ret.Get(0).(func() error); ok {
+		r0 = rf()
+	} else {
+		r0 = ret.Error(0)
+	}
+
+	return r0
+}
+
 // Singleton provides a mock function with given fields: abstract, concrete
 func (_m *Application) Singleton(abstract string, concrete di.BindingFunc) {
 	_m.Called(abstract, concrete)